requirements.txt
fastapi==0.95.2
python-constraint
uvicorn[standard]